Started wrapping security-telemetry components in Coordinator class
